move drawing to its own module
